Fix nos names, alguns métodos

+++ Aerocontrol/app/src/main/java/amsi/dei/estg/ipleiria/aerocontrol/ui/views/MainActivity.java
@@ -53,7 +53,7 @@ public class MainActivity extends AppCompatActivity {
     public void showRestaurantDetails(final int id){
         Restaurant restaurant = SingletonEnterprises.getInstance(this).getRestaurantById(id);
         Intent intent = new Intent(this, RestaurantDetailsActivity.class);
-        intent.putExtra(RestaurantDetailsActivity.RESTAURANT_ID,(int) restaurant.getId());
+        intent.putExtra(RestaurantDetailsActivity.RESTAURANT_ID, restaurant.getId());
         startActivity(intent);
     }
 }
